[apirambo] Move MemoryAccess headers to .. and link them
